requirements.txt
fastapi
uvicorn
feedparser
google-generativeai